Organize adapter and model files
- Rename 'server' to 'apis'
- Add overviews for each api .go file
